feat : test 완료 (swagger X)

+++ src/main/java/com/road_journey/road_journey/items/dto/ItemDto.java
@@ -13,7 +13,7 @@ public class ItemDto {
     private String itemName;
     private String category;
     private String description;
-    private int gold;
+    private long gold;
     private boolean isSelected;
     private boolean isOwned;
